+++ application/service/blocking/src/main/kotlin/org/occurrent/application/service/blocking/ExecuteFilterExtensions.kt
@@ -1,5 +1,7 @@
 package org.occurrent.application.service.blocking
 
+import org.occurrent.application.service.ExecuteFilter
+
 import kotlin.jvm.JvmName
 import kotlin.reflect.KClass